+++ modules/ppcp-settings/resources/css/components/screens/overview/_tab-overview.scss
@@ -1,4 +1,4 @@
-.ppcp-r-tab-dashboard-todo {
+.ppcp-r-tab-overview-todo {
 	margin: 0 0 48px 0;
 }
