+++ modules/ppcp-settings/resources/js/Components/Screens/Overview/TabStyling.js
